requirements.txt
# =========================
# Core utilities
# =========================
python-dotenv
requests
numpy
langdetect

# =========================
# LiveKit (MODERN, COMPATIBLE)
# =========================
livekit>=1.0.19,<2
livekit-agents==1.3.9
livekit-plugins-google

# =========================
# Audio / TTS (HF-safe)
# =========================
edge-tts
pydub

# =========================
# Web server
# =========================
flask
flask-cors
gunicorn
